// src/lib/api/models/createFormSchemaDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */
import type { CreateFormSchemaDtoEntityType } from './createFormSchemaDtoEntityType';
import type { FormSchemaFieldDto } from './formSchemaFieldDto';

export interface CreateFormSchemaDto {
  entityType: CreateFormSchemaDtoEntityType;
  propertyTypeId?: string;
  name: string;
  fields?: FormSchemaFieldDto[];
}